+++ packages/node/src/stellar/api.stellar.spec.ts
@@ -45,7 +45,7 @@ describe('StellarApi', () => {
   it('should fetch block', async () => {
     const latestHeight = await stellarApi.getFinalizedBlockHeight();
     const block = (await stellarApi.fetchBlocks([latestHeight]))[0];
-    expect(block.block.sequence).toEqual(latestHeight);
+    expect(block.getHeader().blockHeight).toEqual(latestHeight);
   });
 
   it('should throw on calling connect', async () => {
@@ -94,7 +94,7 @@ describe('StellarApi', () => {
       { type: { toString: () => 'invoke_host_function' } },
     ]);
     await expect((stellarApi as any).fetchAndWrapLedger(100)).rejects.toThrow(
-      /access a ledger that is after the latest ledger number/,
+      /Not Found/,
     );
   });
 });